mysite/requirements.txt
djangorestframework
requests
rest_framework_swagger